Confirmation Email

+++ client/src/redux/reducers/auth.js
@@ -1,10 +1,17 @@
-import { LOADING, REGISTER_SUCCESS, REGISTER_FAIL } from '../types';
+import {
+  LOADING,
+  REGISTER_SUCCESS,
+  REGISTER_FAIL,
+  CONFIRMATION_SUCCESS,
+  CONFIRMATION_FAIL,
+} from '../types';
 
 const initialState = {
   token: localStorage.getItem('token'),
   isAuthenticated: null,
   loading: false,
   user: null,
+  error: null,
 };
 
 export default function (state = initialState, action) {
@@ -12,14 +19,17 @@ export default function (state = initialState, action) {
 
   switch (type) {
     case REGISTER_SUCCESS:
+    case CONFIRMATION_SUCCESS:
       localStorage.setItem('token', payload.token);
       return {
         ...state,
         token: payload.token,
         isAuthenticated: true,
         loading: false,
+        error: null,
       };
     case REGISTER_FAIL:
+    case CONFIRMATION_FAIL:
       localStorage.removeItem('token');
       return {
         ...state,
@@ -27,6 +37,7 @@ export default function (state = initialState, action) {
         isAuthenticated: false,
         loading: false,
         user: null,
+        error: payload,
       };
     case LOADING:
       return {
